Rewriting the s3cmd command
The old implementation (still valid for the sync command) uses a home
made S3 client while this new one uses the `aws/s3` gem, which is pretty
solid and definitely production ready;

The overall code was also improved. Now s3sync looks more like a library
and the commands are mere front ends for this library;

Here's the list of things I still need to port:

 * The `deleteall` command
 * The `location` command
 * All the flag options read from the command line

Which will be done in the next few commits, I hope;

We're also using Gemfile`s to describe our deps and bundler to test the
software during the development.