Migrating to aws-sdk, the official amazon ruby lib
All the current commands were changed to use the official amazon API and
now I'm using the actual S3 to perform the tests;

The `bin/s3cmd` was changed so we could expand a little bit more the
documentation. In the future, it could be broken down in sections, so
the user wouldn't have to see the full doc when trying to use a certain
command;

The old `s3cmd.rb` file was also killed.