Finishing pretty basic `sync` command (local > remote only)
This command takes a local directory as the source and synchronizes it
with a remote directory. The same idea might be applied to do the other
way around, I just wanted to keep things sane for now;

The brand new command takes the `--exclude` param that receives a ruby
regular expression and filters the upload list. Pretty useful for
filtering backup files for example.

  $ s3cmd sync --exclude="\#" ~/Work/code/lib mybucket:code/